+++ src/platform_impl/linux/wayland/window/mod.rs
@@ -448,6 +448,11 @@ impl Window {
     }
 
     #[inline]
+    pub fn drag_resize_window(&self) -> Result<(), ExternalError> {
+        Err(ExternalError::NotSupported(NotSupportedError::new()))
+    }
+
+    #[inline]
     pub fn set_ime_position(&self, position: Position) {
         let scale_factor = self.scale_factor() as f64;
         let position = position.to_logical(scale_factor);