+++ src/platform_impl/macos/window.rs
@@ -613,6 +613,11 @@ impl UnownedWindow {
         Ok(())
     }
 
+    #[inline]
+    pub fn drag_resize_window(&self) -> Result<(), ExternalError> {
+        Err(ExternalError::NotSupported(NotSupportedError::new()))
+    }
+
     pub(crate) fn is_zoomed(&self) -> bool {
         // because `isZoomed` doesn't work if the window's borderless,
         // we make it resizable temporalily.